DDF-3551 Renamed username related variables to indicate user ID
Previously, the username was passed into the schedule data and used to identify user preferences; variables were named to indicate such. However, another change to DDF changed both the identifier used to get the user preferences and the identifier put inside the schedule data; as such, the function of this code is not affected, but I changed the variable names to clarify the use of the user ID instead of a username.